ou yea, busqueda

+++ src/main/java/com/example/carrotstatebackend/controllers/dtos/request/CreateAgentRequest.java
@@ -19,5 +19,5 @@ public class CreateAgentRequest{
     private String email;
 
     @NotNull
-    private Long managerCode;
+    private Integer managerCode;
 }
